rule 파일 추가
