+++ src/main/java/blackjack/rule/CardGameParticipant.java
@@ -1,7 +1,8 @@
-package blackjack;
+package blackjack.rule;
 
 import java.util.List;
 import java.util.Map;
+import blackjack.Participant;
 
 public interface CardGameParticipant {
         void playTurn(String participant, Map.Entry<String, Participant.ParticipantState> participantEntry, Map<String, List<Integer>> Cards);
